+++ executor/issuetest/BUILD.bazel
@@ -17,10 +17,12 @@ go_test(
         "//parser/auth",
         "//parser/charset",
         "//parser/mysql",
+        "//session",
         "//sessionctx/variable",
         "//statistics",
         "//testkit",
         "//util",
+        "//util/memory",
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
         "@com_github_tikv_client_go_v2//tikv",
